initial iceberg table writer
